+++ DemoApp/DemoApp/Views/TMButtonPage.xaml
@@ -15,7 +15,7 @@
       <Label Text="Floating Buttons"></Label>
       <StackLayout Orientation="Vertical"
                    Spacing="10">
-        <controls:CustomButton x:Name="floatingTextWithLeftIconButton"
+        <controls:TMButton x:Name="floatingTextWithLeftIconButton"
                                Title="Floating Button"
                                IsDisabled="{Binding IsDisabled}"
                                IsFloatingButton="True"
@@ -26,7 +26,7 @@
                                Command="{Binding MyCommand}"
                                CommandParameter="Button5" />
 
-        <controls:CustomButton x:Name="floatingTextWithRightIconButton"
+        <controls:TMButton x:Name="floatingTextWithRightIconButton"
                                Title="Floating Button"
                                IsDisabled="{Binding IsDisabled}"
                                IsFloatingButton="True"
@@ -36,7 +36,7 @@
                                RightIconSource="{x:Static constants:ImageConstants.GalleryIcon}"
                                Command="{Binding MyCommand}"
                                CommandParameter="Button5" />
-        <controls:CustomButton x:Name="floatingIconButton"
+        <controls:TMButton x:Name="floatingIconButton"
                                IsFloatingButton="True"
                                ButtonColor="Primary"
                                IsDisabled="{Binding IsDisabled}"
@@ -45,7 +45,7 @@
                                LeftIconSource="{x:Static constants:ImageConstants.GalleryIcon}"
                                Command="{Binding MyCommand}"
                                CommandParameter="Button5" />
-        <controls:CustomButton x:Name="floatingButtonWithBothIcons"
+        <controls:TMButton x:Name="floatingButtonWithBothIcons"
                                IsFloatingButton="True"
                                ButtonColor="Primary"
                                IsDisabled="{Binding IsDisabled}"
@@ -59,7 +59,7 @@
       <Label Text="Buttons with Images" />
       <StackLayout Orientation="Vertical"
                    Spacing="10">
-        <controls:CustomButton x:Name="primaryleftIconButton"
+        <controls:TMButton x:Name="primaryleftIconButton"
                                Title="Primary"
                                HorizontalOptions="{Binding FullWidthAlignment}"
                                ButtonColor="Primary"
@@ -70,7 +70,7 @@
                                Size="{Binding SelectedFontSize}"
                                ButtonStyle="{Binding SelectedButtonStyle}"
                                CommandParameter="Button5" />
-        <controls:CustomButton x:Name="secondaryBothIconButton"
+        <controls:TMButton x:Name="secondaryBothIconButton"
                                Title="Secondary"
                                HorizontalOptions="{Binding FullWidthAlignment}"
                                ButtonColor="Secondary"
@@ -82,7 +82,7 @@
                                Command="{Binding MyCommand}"
                                CommandParameter="Button5" />
         <StackLayout Orientation="Horizontal">
-          <controls:CustomButton x:Name="tertiaryLeftIconButton"
+          <controls:TMButton x:Name="tertiaryLeftIconButton"
                                  Title="Tertiary"
                                  HorizontalOptions="{Binding FullWidthAlignment}"
                                  ButtonColor="Tertiary"
@@ -93,7 +93,7 @@
                                  RightIconSource="{Binding RightIconSource}"
                                  Command="{Binding MyCommand}"
                                  CommandParameter="Button5" />
-          <controls:CustomButton x:Name="dangerRightIconButton"
+          <controls:TMButton x:Name="dangerRightIconButton"
                                  Title="Danger"
                                  HorizontalOptions="{Binding FullWidthAlignment}"
                                  LeftIconSource="{Binding LeftIconSource}"
