+++ DemoApp/DemoApp/Views/TMCardPage.xaml
@@ -4,6 +4,7 @@
              xmlns:constants="clr-namespace:DemoApp.Constant"
              xmlns:Modus="clr-namespace:Trimble.Modus.Components.Controls.TMCard;assembly=Trimble.Modus.Components"
              xmlns:local="clr-namespace:Trimble.Modus.Components;assembly=Trimble.Modus.Components"
+             xmlns:Button="clr-namespace:Trimble.Modus.Components.Controls.Button;assembly=Trimble.Modus.Components"
              x:Class="DemoApp.Views.TMCardPage"
              Title="TMCardPage">
   <ScrollView>
@@ -92,7 +93,7 @@
           <StackLayout BackgroundColor="Transparent"
                        HorizontalOptions="End"
                        Padding="0,10,0,10">
-            <local:TMButton Title="Button"></local:TMButton>
+            <Button:TMButton Title="Button" Size="Default"></Button:TMButton>
           </StackLayout>
         </StackLayout>
       </Modus:TMCard>
@@ -125,7 +126,7 @@
           <!-- Footer -->
           <StackLayout BackgroundColor="Transparent"
                        Padding="0,10,0,10">
-            <Button Text="Button"
+            <Button:TMButton Title="Button"
                     HorizontalOptions="FillAndExpand" />
           </StackLayout>
         </StackLayout>
@@ -174,10 +175,11 @@
           <StackLayout BackgroundColor="Transparent"
                        Orientation="Horizontal"
                        Spacing="15">
-            <local:TMButton Title="Button"
-                            ButtonStyle="Outline"></local:TMButton>
-            <local:TMButton Title="Button"
-                            ButtonStyle="Fill"></local:TMButton>
+            <Button:TMButton Title="Button"
+                            ButtonStyle="Outline"  Size="Default">
+            </Button:TMButton>
+            <Button:TMButton Title="Button"
+                            ButtonStyle="Fill"  Size="Default"></Button:TMButton>
 
           </StackLayout>
         </StackLayout>
